added waits and views
