+++ src/main/java/com/team2/petrolStation/model/views/SimulatorView.java
@@ -1,0 +1,7 @@
+package com.team2.petrolStation.model.views;
+
+/**
+ * Created by callummarriage on 13/06/2018.
+ */
+public interface SimulatorView {
+}
